+++ image.hpp
@@ -12,18 +12,18 @@ using namespace std;
 class Image {
  
     vector<vector<int> > data;
-    int clacc;
+    short clacc;
     vector<vector<int> > integral;
     vector<Feature> featureVector;
 
     public:
-        Image(string fname, int c);
-        Image(vector<vector<int> > &data, int c);
+        Image(string fname, short c);
+        Image(vector<vector<int> > &data, short c);
         void calIntegral();
         void calFeatureVector();
         void initialize();
         vector<vector<int> > &getImageData();
-        int getImageClass ();
+        short getImageClass ();
         vector<vector<int> > &getImageIntegral();
         vector<Feature> &getFeatureVector();
         void calFeatureByLines(int start, int end);
